modify controller routes, fix theatre id type, add admin role

+++ src/main/java/com/bookit/application/persistence/IShowDao.java
@@ -6,7 +6,7 @@ import com.bookit.application.entity.ShowTimeSlot;
 import java.util.List;
 
 public interface IShowDao extends Crud<Show, String>{
-    List<ShowTimeSlot> getBookedSlotsByTheatreId(Long theatreId);
+    List<ShowTimeSlot> getBookedSlotsByTheatreId(Integer theatreId);
     List<Show> findShowsByMovie(Long movieId);
     List<Show> findShowsByTheatre(Integer theatreId);
 }
